requirements.txt
# --- Core Web Framework ---
Flask==3.0.3
flask-cors==4.0.0
gunicorn==23.0.0
python-dotenv==1.0.1

# --- LangChain / Groq / HuggingFace ---
langchain==0.3.7
langchain-groq==0.2.0
langchain-community==0.3.7
langchain-huggingface==0.1.2
langgraph==0.2.39

# --- Vector Search (FAISS) ---
faiss-cpu==1.12.0
sentence-transformers>=2.6.0

# --- Utilities ---
requests==2.32.3
protobuf==4.25.3
numpy==1.26.4
pydantic==2.8.2
typing-extensions==4.12.2
gunicorn
groq>=0.4.1
pydub==0.25.1
ffmpeg-python==0.2.0
gtts